Move LegacyTimelineEventsController to the legacy folder
